refactor(shared): split shared stuff into new file
